DignoseThatAPISystem/GPU Inference/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.3.0

# ML/CV
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0
